Adicionando validacao nos medicos, requisitos: 1, 2 e 3 completos...

+++ app/controllers/doctors_controller.rb
@@ -15,11 +15,13 @@ class DoctorsController < ApplicationController
 
   def create
     @doctor = Doctor.new doctor_params
+    set_crm
+
     if @doctor.save
       flash[:success] = 'Médico cadastrado com sucesso.'
       redirect_to doctors_path
     else
-      flash[:error] = 'Algo deu errado, tente novamente.'
+      flash[:error] = @doctor.errors.full_messages
       render :new
     end
   end
@@ -28,18 +30,25 @@ class DoctorsController < ApplicationController
   end
 
   def update
+    set_crm
+
     if @doctor.update doctor_params
       flash[:success] = 'Médico atualizado com sucesso.'
       redirect_to doctors_path(@doctor)
     else
-      flash[:error] = 'Algo deu errado, tente novamente.'
+      flash[:error] = @doctor.errors.full_messages[1]
       render :edit
     end
   end
 
   def destroy
-    if @doctor.destroy
-      flash[:success] = 'Médico removido com sucesso.'
+    if @doctor.patients.count === 0
+      if @doctor.destroy
+        flash[:success] = 'Médico removido com sucesso.'
+        redirect_to doctors_path
+      end
+    else
+      flash[:error] = 'Não é possível excluir médico com pacientes.'
       redirect_to doctors_path
     end
   end
@@ -47,6 +56,10 @@ class DoctorsController < ApplicationController
   private
     def instantiate
       @doctor = Doctor.find params[:id]
+    end
+
+    def set_crm
+      @doctor.crm_full = "#{@doctor.crm}-#{@doctor.crm_uf.upcase}"
     end
 
     def doctor_params
